Change importing all things from module from path-end `.*` to names section `> *`
